refactoring  - separate folders for pages
