+++ src/pages/ingredient-page/ingredient-page.jsx
@@ -1,4 +1,4 @@
-import IngridientDetails from '../components/ingredient-details/ingredient-details';
+import IngridientDetails from '../../components/ingredient-details/ingredient-details';
 
 
 export default function IngredientPage() {
