+++ src/pages/order-page/order-page.module.css
@@ -1,0 +1,6 @@
+.center {
+    display: flex;
+    align-items: center;
+    justify-content: center;
+    min-width: 1300px;
+}
