+++ src/pages/reset-password-page/reset-password-page.module.css
@@ -1,0 +1,7 @@
+.container {
+    height: calc(100vh - 64px);
+    display: grid;
+    align-items: center;
+    /* justify-content: center; */
+    text-align: center;
+}
